+++ src/components/DropdownItem/DropdownItem.js
@@ -2,25 +2,15 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
-const Wrapper = styled.div`
-  padding: 0.5rem 1rem;
-  color: ${p => p.theme.grey01};
-  cursor: pointer;
-
-  &:hover {
-    background: ${p => p.theme.grey05};
-  }
-`;
-
 /**
  * Single item in the dropdown panel. This adds the necessary padding for simple text labels.
  *
  * @visibleName Dropdown.Item
  */
 const DropdownItem = ({ children, onClick, ...otherProps }) => (
-  <Wrapper {...otherProps} onClick={onClick}>
+  <div {...otherProps} onClick={onClick} tabIndex="0" role="button">
     {children}
-  </Wrapper>
+  </div>
 );
 
 DropdownItem.propTypes = {
@@ -33,4 +23,9 @@ DropdownItem.defaultProps = {
   onClick: () => null,
 };
 
-export default DropdownItem;
+export default styled(DropdownItem)`
+  padding: 0.5rem 1rem;
+  color: ${p => p.theme.grey01};
+  cursor: pointer;
+  outline: 0px none;
+`;