+++ src/helpers/keyCodes.js
@@ -1,6 +1,12 @@
 const KEY_CODES = {
   Space: 32,
   Enter: 13,
+  Esc: 27,
+  Tab: 8,
+  Left: 37,
+  Up: 38,
+  Right: 39,
+  Down: 40,
 };
 
 export default KEY_CODES;