+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Changed
+- Limit Dropdown max height to 320px by default.
+- Dropdown keyboard navigation support.
+
 ## [3.0.2] - 2018-11-26
 
 ### Fixed